Refactor User Service.
