+++ src/Service/User/UserService.php
@@ -5,31 +5,10 @@ declare(strict_types=1);
 namespace App\Service\User;
 
 use App\Exception\User;
-use App\Repository\UserRepository;
-use App\Service\BaseService;
-use App\Service\RedisService;
 use Firebase\JWT\JWT;
 
-final class UserService extends BaseService
+final class UserService extends Base
 {
-    private const REDIS_KEY = 'user:%s';
-
-    /**
-     * @var UserRepository
-     */
-    protected $userRepository;
-
-    /**
-     * @var RedisService
-     */
-    protected $redisService;
-
-    public function __construct(UserRepository $userRepository, RedisService $redisService)
-    {
-        $this->userRepository = $userRepository;
-        $this->redisService = $redisService;
-    }
-
     public function getAll(): array
     {
         return $this->userRepository->getAll();
@@ -46,38 +25,9 @@ final class UserService extends BaseService
         return $user;
     }
 
-    public function getUserFromCache(int $userId)
-    {
-        $redisKey = sprintf(self::REDIS_KEY, $userId);
-        $key = $this->redisService->generateKey($redisKey);
-        if ($this->redisService->exists($key)) {
-            $data = $this->redisService->get($key);
-            $user = json_decode(json_encode($data), false);
-        } else {
-            $user = $this->getUserFromDb($userId);
-            $this->redisService->setex($key, $user);
-        }
-
-        return $user;
-    }
-
     public function search(string $usersName): array
     {
         return $this->userRepository->search($usersName);
-    }
-
-    public function saveInCache($id, $user): void
-    {
-        $redisKey = sprintf(self::REDIS_KEY, $id);
-        $key = $this->redisService->generateKey($redisKey);
-        $this->redisService->setex($key, $user);
-    }
-
-    public function deleteFromCache($userId): void
-    {
-        $redisKey = sprintf(self::REDIS_KEY, $userId);
-        $key = $this->redisService->generateKey($redisKey);
-        $this->redisService->del($key);
     }
 
     public function create($input)
@@ -158,10 +108,5 @@ final class UserService extends BaseService
         ];
 
         return JWT::encode($token, getenv('SECRET_KEY'));
-    }
-
-    protected function getUserFromDb(int $userId)
-    {
-        return $this->userRepository->getUser($userId);
     }
 }
